requirements.txt
qrcode
Pillow
tqdm
zbarlight
opencv-python
moviepy